// src/shared/types/library.ts
export type LibrarySort =
  | 'default'
  | 'createdAsc'
  | 'createdDesc'
  | 'yearAsc'
  | 'yearDesc'
  | 'titleAsc'
  | 'titleDesc'
  | 'trackCountAsc'
  | 'trackCountDesc'
  | 'albumCountDesc'
  | 'durationAsc'
  | 'durationDesc'
  | 'fileModifiedAsc'
  | 'fileModifiedDesc'
  | 'qualityAsc'
  | 'qualityDesc'
  | 'codecAsc'
  | 'codecDesc'
  | 'audioSpecAsc'
  | 'audioSpecDesc'
  | 'bitrateAsc'
  | 'bitrateDesc'
  | 'bpmAsc'
  | 'bpmDesc'
  | 'trackNumber'
  | 'frequent'
  | 'playCountAsc'
  | 'playCountDesc'
  | 'lastPlayed'
  | 'random'
  | 'title'
  | 'artist'
  | 'artistAlbum'
  | 'album'
  | 'recent';